Add empty item to `ignore-words-list` codespell configuration key
Although rare, false positives from the codespell spell checker tool are virtually inevitable. In order to ease the pain
for whoever has the misfortune to introduce one, I like to pre-populate the .codespellrc configuration file with an
empty `ignore-words-list` configuration key along with an explanatory comment so that the only thing needed is to add
the problematic word.

If that key is left completely empty, it causes codespell to fail:

```
codespell: error: argument -L/--ignore-words-list: expected one argument
```

The workaround is to add an empty item to this comma-separated list, which has no effect on codespell other than to
allow it to run.